Address PR feedback: reorganize MCP server code structure
- Remove emojis from logging and flag descriptions in server.go
- Reorganize MCP server package structure:
  - Split handler.go into individual tool files:
    - search_registry.go: SearchRegistry tool and types
    - run_server.go: RunServer tool and helper functions
    - list_servers.go: ListServers tool and types
    - stop_server.go: StopServer tool and types
    - remove_server.go: RemoveServer tool and types
    - get_server_logs.go: GetServerLogs tool and types
  - Updated handler.go to only contain Handler type and constructor
  - Removed types.go (not needed)
- Fixed linter issues:
  - Renamed ServerInfo to Info to avoid stuttering
  - Fixed formatting issues
- All tests pass and linter is clean
